Add PageHeader component since it's identical across different pages
